fix: fixed bugs with loading of funtions
