+++ src/Config/config.js
@@ -4,5 +4,12 @@ export default {
   type: Phaser.AUTO,
   parent: 'phaser-example',
   width: 800,
-  height: 600
+  height: 600,
+  physics: {
+    default: 'arcade',
+    arcade: {
+        gravity: { y: 300 },
+        debug: false
+    }
+  }
 };
