+++ src/assets/brawlbot_abdomen.png
