+++ src/assets/brawlbot_arm_fist_l.png
