+++ src/assets/brawlbot_chest.png
