+++ src/assets/brawlbot_collar.png
