+++ src/assets/invisible_wall.png
