+++ src/assets/platform.png
