+++ src/assets/skill-desc_0000_foreground.png
