+++ src/assets/skill-desc_0001_buildings.png
