+++ src/assets/skill-desc_0002_far-buildings.png
