+++ src/assets/skill-desc_0003_bg.png
