+++ src/objects/createStars.js
